1.2 DIT Internal Version 1
This version is the first version and has the basic structure of what the final code is kind off going to look like. In this version there are 3 different options for the user to return data from the database. The database contains information that can reveal insights into the popularity of songs.