1.2 DIT Internal V2
Version 2 of the MusicData code. Added functions like the help function and ask again function in order to make the users experience easier and more fluid. Added column search function so that the user can search through the data by column, making the user be able to search exactly what they want rather than using fixed  SQL statements. Added menu function makes the program more simple, it also contains 2 more options to select from the menu. This function  and the ask again function contain error catching using a try and except, making the program more robust. The database was also edited from 150 songs in it to the top 59 songs based on what songs had the most streams. This makes the output for the user clearer as they are no longer overwhelmed with the large amount of data.