1.2 DIT Internal Version 3
Completely restructured column search function, meaning that the user can search for more than one column. Also made it more robust. Added a list of the column names so that when the user inputed a column name in the column search function the program can check if the input is correct or not. Edited the help command so that it was more useful for the V3 program. Minor bug fixes like in the ask again function, this being done by adding an exit rather than a break to completely stop the code and prevent it for going into another iteration of the menu while loop. Added clear console option in the menu so the user can remove previous data from the console if they want to.