1.2 DIT Internal Version 4
Many features have been added to this version. These include the search by artist function so the user can search through the database, returning all songs and their data associated with the artist. The search by song name function which lets the user search through the database by inputting the name of the song they want to return. The update streams function which lets the user select a song and update the amount of streams that it has as this may have changed since the time that the streams were recorded and inputted into the database.  The delete data function so the user can remove songs and their data that they do not want to see turn up in their queries. Minor bug and formatting changes for a better user experience. Minor changes to the help function for a better user experience.